Live location sharing - extract live time UI for reuse (#8283)
* extract livetimeremaining into own component


* extract LiveTimeRemaining for reuse in beacon timeline
