+++ res/css/components/views/beacon/_RoomLiveShareWarning.scss
@@ -39,12 +39,6 @@ limitations under the License.
     font-size: $font-15px;
 }
 
-.mx_RoomLiveShareWarning_expiry {
-    color: $secondary-content;
-    font-size: $font-12px;
-    margin-right: $spacing-16;
-}
-
 .mx_RoomLiveShareWarning_spinner {
     margin-right: $spacing-16;
 }
